change type of event id to intptr_t for better hardware compatibility
for 32 bit platforms int16_t is not capable of holding addres pointer
value.

+++ utility/EventBase.cpp
@@ -3,7 +3,7 @@
 EventBase::EventBase(void)
 {
     // use object's address as event id
-    id = (uint16_t) this;
+    id = (intptr_t) this;
     
     repeatCount = 0;
     period = 0;
